Refactor middlewares
